refactor(projects): :art: Move projects into a projects directory to help keep the structure clearer
