+++ .projen/files.json
@@ -14,9 +14,9 @@
     "CODEOWNERS",
     "LICENSE",
     "README.md",
-    "src/jsii/jsii-project-options.ts",
-    "src/npm/npm-package-options.ts",
-    "src/python/python-package-options.ts",
+    "src/projects/jsii/jsii-project-options.ts",
+    "src/projects/npm/npm-package-options.ts",
+    "src/projects/python/python-package-options.ts",
     "tsconfig.dev.json"
   ],
   "//": "~~ Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\"."
