+++ src/index.ts
@@ -1,4 +1,2 @@
-export * from './jsii';
-export * from './npm';
-export * from './python';
 export * from './components';
+export * from './projects';
