+++ src/projects/jsii/jsii-project-options.ts
@@ -1,6 +1,6 @@
 // ~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".
 import type { cdk, github, GitOptions, GroupRunnerOptions, IgnoreFileOptions, javascript, LoggerOptions, Project, ProjectType, ProjenrcJsonOptions, ReleasableCommits, release, RenovatebotOptions, typescript } from 'projen';
-import type { ReadmeOptions } from '../';
+import type { ReadmeOptions } from '../../';
 
 /**
  * JsiiProjectOptions
