+++ src/projects/python/python-package-options.ts
@@ -1,6 +1,6 @@
 // ~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".
 import type { github, GitOptions, IgnoreFileOptions, javascript, LoggerOptions, Project, ProjectType, ProjenrcJsonOptions, python, RenovatebotOptions, typescript } from 'projen';
-import type { ReadmeOptions } from '../';
+import type { ReadmeOptions } from '../../';
 
 /**
  * PythonPackageOptions
