+++ test/projects/npm/index.test.ts
@@ -1,15 +1,11 @@
 import { synthSnapshot } from 'projen/lib/util/synth';
-import { PythonPackage } from '../../src/python';
+import { NpmPackage } from '../../../src/projects/npm';
 
-describe('Python Package', () => {
+describe('NPM Package', () => {
   it('synthesizes', () => {
-    const project = new PythonPackage({
-      authorEmail: 'test@example.com',
-      authorName: 'test',
+    const project = new NpmPackage({
       codeOwners: ['test'],
-      moduleName: 'test',
       name: 'test',
-      version: '0.1.0',
     });
 
     const output = synthSnapshot(project);
