+++ test/projects/python/index.test.ts
@@ -1,11 +1,15 @@
 import { synthSnapshot } from 'projen/lib/util/synth';
-import { NpmPackage } from '../../src/npm';
+import { PythonPackage } from '../../../src/projects/python';
 
-describe('NPM Package', () => {
+describe('Python Package', () => {
   it('synthesizes', () => {
-    const project = new NpmPackage({
+    const project = new PythonPackage({
+      authorEmail: 'test@example.com',
+      authorName: 'test',
       codeOwners: ['test'],
+      moduleName: 'test',
       name: 'test',
+      version: '0.1.0',
     });
 
     const output = synthSnapshot(project);
